Obstacle-handling in fill-1
Goal is to add a list of indecies where there will be obstacles and then filling the board with patterns without overwritting the obstacles, or -1.
Due to the logic behind fill-1 and fill-2, it was easier to add this feature to fill-1 so that was done. Need more time to plan out how to add this with fill-2.