working on addTurn
Was unsure if minTurn and addTurn should be methods of TesterBoard or make new class of TesterPathBoard, but since I need a helper function of addTurn that will extend 1 on (hidden) board in all 4 directions, it seems like that shouldn't be method of Board or Gameboard. Thus I will attempt to make new TesterPathBoard. Saving progress made so far so I can come back to it if  TesterPathBoard crashes and burns...