Finds least amount of turns needed
Newly made TesterPathBoard is working nicely. Completed minTurns(); it will return a number outside of maxTurns allowed if A' can't be reached within the current max which is 3.
Notes are left in PearTester on how to make addTurn more efficient, and will also make a more intuitive paths board when it's called.