Updated effects on removed pairs
Fixed an error where pairs that have been removed already can still appear to be clicked on. Also, the match function will no longer be called when either of the grid elements player has clicked on has value of 0.
Better font styles and sizes were added to texts and integers on the board to increase user legibility.