Fixed end of game error and minor appearance changes
Previously, when a player finished all the levels, the game board would be cleared, but an error would occur in console. This is because when the last level is over, the program would try to create new level by clearing the game board, but get stuck as there are no more PLevel's in the game array to create next level from. This was fixed by calling and checking for existence of anymore levels and calling endOfGame() instead.

Additionally, a few changes to font and colours was made.